+++ app/src/main/java/com/ljmaq/budgetrule/features/record/data/repository/partition/InvestmentsRepositoryImpl.kt
@@ -1,11 +1,8 @@
 package com.ljmaq.budgetrule.features.record.data.repository.partition
 
 import com.ljmaq.budgetrule.features.record.data.datasource.partition.investments.InvestmentsDao
-import com.ljmaq.budgetrule.features.record.data.datasource.partition.needs.NeedsDao
 import com.ljmaq.budgetrule.features.record.domain.model.partition.Investments
-import com.ljmaq.budgetrule.features.record.domain.model.partition.Needs
 import com.ljmaq.budgetrule.features.record.domain.repository.partition.InvestmentsRepository
-import com.ljmaq.budgetrule.features.record.domain.repository.partition.NeedsRepository
 import kotlinx.coroutines.flow.Flow
 
 class InvestmentsRepositoryImpl(
@@ -15,7 +12,7 @@ class InvestmentsRepositoryImpl(
         return dao.getInvestments()
     }
 
-    override suspend fun getInvestmentsById(id: Int): Investments? {
+    override suspend fun getInvestmentsById(id: Long): Investments? {
         return dao.getInvestmentsById(id)
     }
