+++ app/src/main/java/com/ljmaq/budgetrule/features/record/data/repository/partition/NeedsRepositoryImpl.kt
@@ -12,7 +12,7 @@ class NeedsRepositoryImpl(
         return dao.getNeeds()
     }
 
-    override suspend fun getNeedsById(id: Int): Needs? {
+    override suspend fun getNeedsById(id: Long): Needs? {
         return dao.getNeedsById(id)
     }
