+++ app/src/main/java/com/ljmaq/budgetrule/features/record/domain/repository/partition/InvestmentsRepository.kt
@@ -1,13 +1,12 @@
 package com.ljmaq.budgetrule.features.record.domain.repository.partition
 
 import com.ljmaq.budgetrule.features.record.domain.model.partition.Investments
-import com.ljmaq.budgetrule.features.record.domain.model.partition.Needs
 import kotlinx.coroutines.flow.Flow
 
 interface InvestmentsRepository {
     fun getInvestments(): Flow<List<Investments>>
 
-    suspend fun getInvestmentsById(id: Int): Investments?
+    suspend fun getInvestmentsById(id: Long): Investments?
 
     suspend fun insertInvestments(investments: Investments)
