+++ app/src/main/java/com/ljmaq/budgetrule/features/record/domain/repository/partition/NeedsRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface NeedsRepository {
     fun getNeeds(): Flow<List<Needs>>
 
-    suspend fun getNeedsById(id: Int): Needs?
+    suspend fun getNeedsById(id: Long): Needs?
 
     suspend fun insertNeeds(needs: Needs)
