+++ app/src/main/java/com/ljmaq/budgetrule/features/record/domain/repository/partition/SavingsRepository.kt
@@ -1,13 +1,12 @@
 package com.ljmaq.budgetrule.features.record.domain.repository.partition
 
-import com.ljmaq.budgetrule.features.record.domain.model.partition.Needs
 import com.ljmaq.budgetrule.features.record.domain.model.partition.Savings
 import kotlinx.coroutines.flow.Flow
 
 interface SavingsRepository {
     fun getSavings(): Flow<List<Savings>>
 
-    suspend fun getSavingsById(id: Int): Savings?
+    suspend fun getSavingsById(id: Long): Savings?
 
     suspend fun insertSavings(savings: Savings)
